SM120: move triton routing helper to light module so validator can test it
